requirements.txt
streamlit
requests
feedparser
langchain
groq
python-dotenv
rouge-score
